requirements.txt
fastapi==0.115.5
mysql_connector_repackaged==0.3.1
pydantic==2.10.2
Requests==2.32.3
